Consolidated the three divergent model definitions into shared src/models.py
users/servers/command_usage were defined independently in bot.py,
stripe_webhook_service.py, and subscription_manager.py with
disagreeing column lengths, nullability, and columns (e.g. discord_id
String(30) vs String(50); bot's servers had minimum_age and the
instructions columns while the manager's had email/last_renewal_date
and neither had both) -- whichever service started first won the
create_all race and defined the real schema.

src/models.py is now the single source of truth: one engine, one
Session factory, one session_scope, and one definition of each model
(the union of all columns). All four services import from it; the
duplicate model classes, engines, and session_scope copies are
deleted. create_all no longer runs at service import -- models.py
exposes init_db() for tests, and real schema management moves to
Alembic in the next commit.

conftest.py creates the tables on the shared in-memory test engine and
pins a single module identity for models.py (tests import services
both as 'x' and 'src.x', which would otherwise produce two module
objects with two separate in-memory databases). The engine safety-net
check now includes models itself. Updated the two tests that
referenced the removed SessionVerification.
