+++ src/bot.py
@@ -1,21 +1,21 @@
-import hashlib
 import os
 import json
 import asyncio
 import logging
 import time
 from datetime import datetime, timedelta, timezone
-from contextlib import contextmanager
 
 import discord
 from discord import app_commands
 from dotenv import load_dotenv
 import pika
 import stripe
-from sqlalchemy import create_engine, Column, Integer, String, Boolean, DateTime, text, inspect
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
 from cryptography.fernet import Fernet
+
+try:
+    from .models import User, Server, CommandUsage, session_scope
+except ImportError:
+    from models import User, Server, CommandUsage, session_scope
 
 # Load environment variables
 load_dotenv()
@@ -36,7 +36,6 @@ logging.getLogger('pika').setLevel(getattr(logging, PIKA_LOG_LEVEL, logging.WARN
 # Retrieve environment variables
 DISCORD_BOT_TOKEN = os.getenv('DISCORD_BOT_TOKEN')
 STRIPE_SECRET_KEY = os.getenv('STRIPE_SECRET_KEY')
-DATABASE_URL = os.getenv('DATABASE_URL_VERIFICATION')
 
 
 # RabbitMQ Configuration
@@ -53,11 +52,6 @@ required_env_vars = ['DISCORD_BOT_TOKEN', 'STRIPE_SECRET_KEY', 'DATABASE_URL_VER
 missing_vars = [var for var in required_env_vars if not os.getenv(var)]
 if missing_vars:
     raise EnvironmentError(f"Missing required environment variables: {', '.join(missing_vars)}")
-
-# Database setup
-engine = create_engine(DATABASE_URL)
-Base = declarative_base()
-Session = sessionmaker(bind=engine)
 
 # Initialize the Discord bot with intents (avoid heavy member chunking at startup)
 intents = discord.Intents.default()
@@ -191,49 +185,6 @@ def decrypt_dob(encrypted_dob: str) -> datetime:
     dob_str = dob_bytes.decode('utf-8')  # Convert bytes back to string
     return datetime.strptime(dob_str, '%Y-%m-%d')  # Convert string to datetime object
 
-# Modify User model to store the encrypted DOB
-class User(Base):
-    __tablename__ = 'users'
-    id = Column(Integer, primary_key=True)
-    discord_id = Column(String(30), nullable=False)
-    verification_status = Column(Boolean, default=False)
-    last_verification_attempt = Column(DateTime(timezone=True))
-    dob = Column(String(255), nullable=True)  # Store encrypted DOB
-
-    @staticmethod
-    def get_current_time():
-        return datetime.now(timezone.utc)
-
-    def set_verification_attempt(self):
-        self.last_verification_attempt = self.get_current_time()
-
-class Server(Base):
-    __tablename__ = 'servers'
-    id = Column(Integer, primary_key=True)
-    server_id = Column(String, unique=True, nullable=False)
-    owner_id = Column(String, nullable=False)
-    role_id = Column(String, nullable=True)
-    tier = Column(String, nullable=False)
-    subscription_status = Column(Boolean, default=False)
-    verifications_count = Column(Integer, default=0)
-    subscription_start_date = Column(DateTime, nullable=True)
-    stripe_subscription_id = Column(String, nullable=True)
-    minimum_age = Column(Integer, nullable=False, default=18)
-    instructions_channel_id = Column(String, nullable=True)
-    instructions_message_id = Column(String, nullable=True)
-
-
-class CommandUsage(Base):
-    __tablename__ = 'command_usage'
-    id = Column(Integer, primary_key=True)
-    server_id = Column(String(30), nullable=False)
-    user_id = Column(String(30), nullable=False)
-    command = Column(String(50), nullable=False)
-    timestamp = Column(DateTime(timezone=True), nullable=False)
-
-# Create tables
-Base.metadata.create_all(engine)
-
 # RabbitMQ setup
 credentials = pika.PlainCredentials(RABBITMQ_USERNAME, RABBITMQ_PASSWORD)
 
@@ -283,18 +234,6 @@ def get_rabbitmq_channel():
     channel = connection.channel()
     channel.queue_declare(queue=RABBITMQ_QUEUE_NAME, durable=True)
     return channel
-
-@contextmanager
-def session_scope():
-    session = Session()
-    try:
-        yield session
-        session.commit()
-    except:
-        session.rollback()
-        raise
-    finally:
-        session.close()
 
 def get_server_config(guild_id):
     with session_scope() as session:
