+++ tests/test_subscription_checker.py
@@ -8,7 +8,9 @@ import subscription_checker
 
 
 def test_check_subscriptions_queries_verification_db():
-    with patch.object(subscription_checker, "SessionVerification") as mock_session_v:
+    with patch.object(subscription_checker, "session_scope") as mock_scope:
+        mock_session = MagicMock()
+        mock_scope.return_value.__enter__.return_value = mock_session
         subscription_checker.check_subscriptions()
-        mock_session_v.assert_called()
-        mock_session_v.return_value.query.assert_called()
+        mock_scope.assert_called()
+        mock_session.query.assert_called()
